requirements.txt
fastapi
uvicorn
python-multipart
opencv-python
torch
clip  # or use: git+https://github.com/openai/CLIP.git
Pillow
whisper
transformers
kafka-python
redis
ffmpeg-python
streamlit
requests
